Handle UnicodeError in SubProcess
Using vscode on a windows 11 computer, I noticced that there is a very rare case where
there is a non unicode character that is passed through.

Specifying subprocess text as True will through a UnicdodeDecodeError.

To get around it we decode each line to utf-8 and replace any non unicode characters that may occur  with a placeholder character (�).